tests/wpt/web-platform-tests/infrastructure/testdriver/actions/eventOrder.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver actions: event order</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<button id="a">Button a</button>
<button id="b">Button b</button>

<script>
// Pointer 1 is added before Pointer 2 so it comes first in the list of sources
// Therefore its actions happen first
let events = [];

async_test(t => {
  Array.prototype.forEach.call(document.getElementsByTagName("button"),
                               (x) => x.addEventListener("mousedown", () => {events.push(x.id)}));

  let button_a = document.getElementById("a");
  let button_b = document.getElementById("b");
  let actions = new test_driver.Actions()
    .addPointer("pointer1")
    .addPointer("pointer2")
    .pointerMove(0, 0, {origin: button_a, sourceName: "pointer1"})
    .pointerMove(0, 0, {origin: button_b, sourceName: "pointer2"})
    .pointerDown({sourceName: "pointer2"})
    .pointerDown({sourceName: "pointer1"})
    .pointerUp({sourceName: "pointer2"})
    .pointerUp({sourceName: "pointer1"})
    .send()
    .then(t.step_func_done(() => assert_array_equals(events, ["a", "b"])))
    .catch(e => t.step_func(() => assert_unreached("Actions sequence failed " + e)));
});
</script>
